Started the long refactor process.
