+++ server/deckr/core/__init__.py
@@ -1,3 +1,7 @@
 """
 This package contains all of the core functions and services that don't belong in any other package.
+This includes:
+    * The GameMaster and its associated functionality.
+    * The service architecture
+        * The reloading service wrapper is found in debug.
 """
